+++ source/ui.py
@@ -50,7 +50,7 @@ class TEXTUREBAKE_PT_main(bpy.types.Panel):
         row = layout.row()
         row.prop(context.scene.TextureBake_Props, "background_bake", expand = True)
 
-        if bpy.context.scene.TextureBake_Props.background_bake == "bg":
+        if context.scene.TextureBake_Props.background_bake == "bg":
             layout.row().prop(context.scene.TextureBake_Props, "background_bake_name", text="Name")
 
         row = layout.row()
@@ -63,7 +63,7 @@ class TEXTUREBAKE_PT_main(bpy.types.Panel):
 
         if len(background_bake_ops.bgops_list) == 0 and len(background_bake_ops.bgops_list_finished) == 0:
             row = box.row()
-            row.label(text="No running or finished background bakes", icon="MONKEY")
+            row.label(text="No running or finished background bakes", icon='MONKEY')
         else:
             for p in background_bake_ops.bgops_list:
                 t = Path(tempfile.gettempdir())
@@ -80,7 +80,7 @@ class TEXTUREBAKE_PT_main(bpy.types.Panel):
                 name = p[3]
                 if name == "":
                     name = "Untitled"
-                row.label(text=f"{name} - baking in progress {progress}%", icon="GHOST_DISABLED")
+                row.label(text=f"{name} - baking in progress {progress}%", icon='GHOST_DISABLED')
 
         if len(background_bake_ops.bgops_list_finished) != 0:
             for p in background_bake_ops.bgops_list_finished:
@@ -89,15 +89,15 @@ class TEXTUREBAKE_PT_main(bpy.types.Panel):
                 name = p[3]
                 if name == "":
                     name = "Untitled"
-                col.label(text=f"{name} - finished!", icon="GHOST_ENABLED")
+                col.label(text=f"{name} - finished!", icon='GHOST_ENABLED')
                 col = row.column()
-                col.operator("texture_bake.bake_import_individual", text="", icon="IMPORT").pnum = int(p[0].pid)
+                col.operator("texture_bake.bake_import_individual", text="", icon='IMPORT').pnum = int(p[0].pid)
                 col = row.column()
-                col.operator("texture_bake.bake_delete_individual", text="", icon="CANCEL").pnum = int(p[0].pid)
+                col.operator("texture_bake.bake_delete_individual", text="", icon='CANCEL').pnum = int(p[0].pid)
 
         row = box.row()
-        row.operator("texture_bake.bake_import", text="Import all", icon="IMPORT")
-        row.operator("texture_bake.bake_delete", text="Discard all", icon="TRASH")
+        row.operator("texture_bake.bake_import", text="Import all", icon='IMPORT')
+        row.operator("texture_bake.bake_delete", text="Discard all", icon='TRASH')
         row.enabled = len(background_bake_ops.bgops_list_finished) != 0
 
 
@@ -114,13 +114,13 @@ class TEXTUREBAKE_PT_presets(TextureBakeCategoryPanel, bpy.types.Panel):
                           "presets_list", context.scene.TextureBake_Props, "presets_list_index")
 
         col = row.column()
-        col.operator("texture_bake.refresh_presets", text="", icon="FILE_REFRESH")
-        col.operator("texture_bake.load_preset", text="", icon="CHECKMARK")
-        col.operator("texture_bake.delete_preset", text="", icon="CANCEL")
+        col.operator("texture_bake.refresh_presets", text="", icon='FILE_REFRESH')
+        col.operator("texture_bake.load_preset", text="", icon='CHECKMARK')
+        col.operator("texture_bake.delete_preset", text="", icon='CANCEL')
 
         row = layout.row()
         row.prop(context.scene.TextureBake_Props, "preset_name")
-        row.operator("texture_bake.save_preset", text="", icon="ADD")
+        row.operator("texture_bake.save_preset", text="", icon='ADD')
 
 
 class TEXTUREBAKE_PT_objects(TextureBakeCategoryPanel, bpy.types.Panel):
@@ -138,15 +138,15 @@ class TEXTUREBAKE_PT_objects(TextureBakeCategoryPanel, bpy.types.Panel):
             col.template_list("TEXTUREBAKE_UL_object_list", "", context.scene.TextureBake_Props,
                             "object_list", context.scene.TextureBake_Props, "object_list_index")
             col = row.column()
-            col.operator('texture_bake.add_object', text="", icon="ADD")
-            col.operator('texture_bake.remove_object', text="", icon="REMOVE")
-            col.operator('texture_bake.clear_objects', text="", icon="TRASH")
+            col.operator('texture_bake.add_object', text="", icon='ADD')
+            col.operator('texture_bake.remove_object', text="", icon='REMOVE')
+            col.operator('texture_bake.clear_objects', text="", icon='TRASH')
             col.separator()
-            col.operator('texture_bake.move_object', text="", icon="TRIA_UP").direction="UP"
-            col.operator('texture_bake.move_object', text="", icon="TRIA_DOWN").direction="DOWN"
+            col.operator('texture_bake.move_object', text="", icon='TRIA_UP').direction="UP"
+            col.operator('texture_bake.move_object', text="", icon='TRIA_DOWN').direction="DOWN"
 
         layout.row().prop(context.scene.TextureBake_Props, "selected_to_target")
-        if bpy.context.scene.TextureBake_Props.selected_to_target:
+        if context.scene.TextureBake_Props.selected_to_target:
             layout.use_property_split = True
             layout.row().prop(context.scene.TextureBake_Props, "target_object")
             layout.row().prop(context.scene.render.bake, "cage_object", text="Cage Object (Optional)")
@@ -174,18 +174,18 @@ class TEXTUREBAKE_PT_input(TextureBakeCategoryPanel, bpy.types.Panel):
         row.prop(context.scene.TextureBake_Props, "selected_curvature")
         row.prop(context.scene.TextureBake_Props, "selected_lightmap")
 
-        if bpy.context.scene.TextureBake_Props.selected_lightmap and bpy.context.scene.TextureBake_Props.global_mode == "pbr_bake":
+        if context.scene.TextureBake_Props.selected_lightmap and context.scene.TextureBake_Props.global_mode == "pbr_bake":
             layout.row().prop(context.scene.cycles, "samples")
 
-        if bpy.context.scene.TextureBake_Props.selected_lightmap:
+        if context.scene.TextureBake_Props.selected_lightmap:
             row = layout.row()
             row.prop(context.scene.TextureBake_Props, "selected_lightmap_denoise")
-            if not bpy.context.scene.TextureBake_Props.export_textures:
+            if not context.scene.TextureBake_Props.export_textures:
                 row.enabled = False
 
             row = layout.row()
             row.prop(context.scene.TextureBake_Props, "lightmap_apply_colman")
-            if not bpy.context.scene.TextureBake_Props.export_textures:
+            if not context.scene.TextureBake_Props.export_textures:
                 row.enabled = False
 
         layout.row().operator("texture_bake.import_materials", icon='ADD')
@@ -248,10 +248,10 @@ class TEXTUREBAKE_PT_output(TextureBakeCategoryPanel, bpy.types.Panel):
         if(context.scene.TextureBake_Props.global_mode == "cycles_bake"):
             layout.use_property_split = True
 
-            cscene = bpy.context.scene.cycles
+            cscene = context.scene.cycles
             layout.row().prop(cscene, "bake_type")
 
-            cbk = bpy.context.scene.render.bake
+            cbk = context.scene.render.bake
             if cscene.bake_type == 'NORMAL':
                 layout.row().prop(cbk, "normal_space", text="Space")
                 layout.row().prop(cbk, "normal_r", text="Swizzle R")
@@ -274,7 +274,7 @@ class TEXTUREBAKE_PT_output(TextureBakeCategoryPanel, bpy.types.Panel):
 
             row = layout.row()
             row.prop(context.scene.TextureBake_Props, "run_denoise", text="Denoise CyclesBake")
-            if not bpy.context.scene.TextureBake_Props.export_textures:
+            if not context.scene.TextureBake_Props.export_textures:
                 row.enabled = False
 
 
@@ -342,7 +342,7 @@ class TEXTUREBAKE_PT_export_settings(TextureBakeCategoryPanel, bpy.types.Panel):
 
                 row = layout.row()
                 row.prop(context.scene.TextureBake_Props, "export_folder_per_object")
-                row.enabled = bpy.context.scene.TextureBake_Props.export_textures
+                row.enabled = context.scene.TextureBake_Props.export_textures
 
                 if context.scene.TextureBake_Props.export_mesh:
                     layout.row().prop(context.scene.TextureBake_Props, "export_apply_modifiers")
@@ -350,7 +350,7 @@ class TEXTUREBAKE_PT_export_settings(TextureBakeCategoryPanel, bpy.types.Panel):
 
                 layout.row().prop(context.scene.TextureBake_Props, "export_datetime")
 
-                if context.scene.TextureBake_Props.export_mesh and not bpy.context.scene.TextureBake_Props.export_folder_per_object:
+                if context.scene.TextureBake_Props.export_mesh and not context.scene.TextureBake_Props.export_folder_per_object:
                     layout.row().prop(context.scene.TextureBake_Props, "fbx_name")
 
                 layout.row().prop(context.scene.TextureBake_Props, "export_file_format", text="Format")
@@ -365,10 +365,10 @@ class TEXTUREBAKE_PT_export_settings(TextureBakeCategoryPanel, bpy.types.Panel):
                 row = layout.row()
                 row.prop(context.scene.TextureBake_Props, "export_color_space")
                 if context.scene.TextureBake_Props.global_mode == "pbr_bake":
-                    row.enabled = bpy.context.scene.TextureBake_Props.selected_col
+                    row.enabled = context.scene.TextureBake_Props.selected_col
 
                 if context.scene.TextureBake_Props.global_mode == "cycles_bake":
-                    row.enabled = bpy.context.scene.cycles.bake_type != "NORMAL"
+                    row.enabled = context.scene.cycles.bake_type != "NORMAL"
         else:
             layout.row().label(text="Unavailable - Blend file not saved")
 
@@ -384,7 +384,7 @@ class TEXTUREBAKE_PT_uv(TextureBakeCategoryPanel, bpy.types.Panel):
         row.prop(context.scene.TextureBake_Props, "uv_mode", expand=True)
         row.enabled = context.scene.TextureBake_Props.export_textures
 
-        if bpy.context.scene.TextureBake_Props.uv_mode == "udims":
+        if context.scene.TextureBake_Props.uv_mode == "udims":
             layout.row().prop(context.scene.TextureBake_Props, "udim_tiles")
 
         row = layout.row()
@@ -405,7 +405,7 @@ class TEXTUREBAKE_PT_other(TextureBakeCategoryPanel, bpy.types.Panel):
         row.prop(context.scene.TextureBake_Props, "batch_name")
 
         row = layout.row()
-        if bpy.context.scene.TextureBake_Props.background_bake == "fg":
+        if context.scene.TextureBake_Props.background_bake == "fg":
             text = "Copy objects and apply bakes"
         else:
             text = "Copy objects and apply bakes (after import)"
@@ -414,7 +414,7 @@ class TEXTUREBAKE_PT_other(TextureBakeCategoryPanel, bpy.types.Panel):
         row.enabled = not context.scene.TextureBake_Props.tex_per_mat
 
         if (context.scene.TextureBake_Props.prep_mesh):
-            if bpy.context.scene.TextureBake_Props.background_bake == "fg":
+            if context.scene.TextureBake_Props.background_bake == "fg":
                 text = "Hide source objects after bake"
             else:
                 text = "Hide source objects after bake (after import)"
@@ -423,18 +423,18 @@ class TEXTUREBAKE_PT_other(TextureBakeCategoryPanel, bpy.types.Panel):
 
             row = layout.row()
             row.prop(context.scene.TextureBake_Props, "create_gltf_node")
-            if bpy.context.scene.TextureBake_Props.create_gltf_node:
+            if context.scene.TextureBake_Props.create_gltf_node:
                 row.prop(context.scene.TextureBake_Props, "gltf_selection", text="")
 
         row = layout.row()
         row.prop(context.scene.TextureBake_Props, "preserve_materials")
 
-        if bpy.context.preferences.addons["cycles"].preferences.has_active_device():
+        if context.preferences.addons["cycles"].preferences.has_active_device():
             row = layout.row()
             row.prop(context.scene.cycles, "device")
             row = layout.row()
             row.prop(context.scene.TextureBake_Props, "memory_limit")
-            row.enabled = bpy.context.scene.cycles.device != "CPU"
+            row.enabled = context.scene.cycles.device != "CPU"
         else:
             row=layout.row()
             row.label(text="No valid GPU device in Blender Preferences. Using CPU.")
@@ -451,7 +451,7 @@ class TEXTUREBAKE_PT_packing(TextureBakeCategoryPanel, bpy.types.Panel):
             layout.row().label(text="Unavailable - Channel packing requires PBR bake mode")
         elif not functions.is_blend_saved():
             layout.row().label(text="Unavailable - Blend file not saved")
-        elif not bpy.context.scene.TextureBake_Props.export_textures:
+        elif not context.scene.TextureBake_Props.export_textures:
             layout.row().label(text="Unavailable - You must be exporting your bakes")
         else:
             row = layout.row()
@@ -459,8 +459,8 @@ class TEXTUREBAKE_PT_packing(TextureBakeCategoryPanel, bpy.types.Panel):
             col.template_list("TEXTUREBAKE_UL_packed_textures", "", context.scene.TextureBake_Props,
                                 "cp_list", context.scene.TextureBake_Props, "cp_list_index")
             col = row.column()
-            col.operator("texture_bake.delete_packed_texture", text="", icon="CANCEL")
-            col.operator("texture_bake.reset_packed_textures", text="", icon="MONKEY")
+            col.operator("texture_bake.delete_packed_texture", text="", icon='CANCEL')
+            col.operator("texture_bake.reset_packed_textures", text="", icon='PRESET_NEW')
 
             layout.row().prop(context.scene.TextureBake_Props, "cp_name")
             layout.row().prop(context.scene.TextureBake_Props, "cp_file_format", text="Format")
@@ -469,21 +469,21 @@ class TEXTUREBAKE_PT_packing(TextureBakeCategoryPanel, bpy.types.Panel):
             layout.row().prop(context.scene.TextureBake_Props, "cptex_B", text="B")
             layout.row().prop(context.scene.TextureBake_Props, "cptex_A", text="A")
 
-            cp_list = bpy.context.scene.TextureBake_Props.cp_list
-            current_name = bpy.context.scene.TextureBake_Props.cp_name
+            cp_list = context.scene.TextureBake_Props.cp_list
+            current_name = context.scene.TextureBake_Props.cp_name
             if current_name in cp_list: # Editing a cpt that is already there
                 index = cp_list.find(current_name)
                 cpt = cp_list[index]
 
-                if (cpt.R != bpy.context.scene.TextureBake_Props.cptex_R
-                    or cpt.G != bpy.context.scene.TextureBake_Props.cptex_G
-                    or cpt.B != bpy.context.scene.TextureBake_Props.cptex_B
-                    or cpt.A != bpy.context.scene.TextureBake_Props.cptex_A
-                    or cpt.file_format != bpy.context.scene.TextureBake_Props.cp_file_format):
+                if (cpt.R != context.scene.TextureBake_Props.cptex_R
+                    or cpt.G != context.scene.TextureBake_Props.cptex_G
+                    or cpt.B != context.scene.TextureBake_Props.cptex_B
+                    or cpt.A != context.scene.TextureBake_Props.cptex_A
+                    or cpt.file_format != context.scene.TextureBake_Props.cp_file_format):
                     row = layout.row()
                     row.alert=True
                     text = f"Update {current_name} (!!not saved!!)"
-                    row.operator("texture_bake.add_packed_texture", text=text, icon="ADD")
+                    row.operator("texture_bake.add_packed_texture", text=text, icon='ADD')
                 else: # No changes, no button
                     text = f"Editing {current_name}"
                     row = layout.row()
@@ -493,7 +493,7 @@ class TEXTUREBAKE_PT_packing(TextureBakeCategoryPanel, bpy.types.Panel):
                 row = layout.row()
                 text = "Add new (!!not saved!!)"
                 row.alert = True
-                row.operator("texture_bake.add_packed_texture", text=text, icon="ADD")
+                row.operator("texture_bake.add_packed_texture", text=text, icon='ADD')
 
 
 class TextureBakePreferences(bpy.types.AddonPreferences):
@@ -681,27 +681,29 @@ class TEXTUREBAKE_OT_add_object(Operator):
     """Add selected object(s) to the bake list"""
     bl_idname = "texture_bake.add_object"
     bl_label = "Add a new object to bake list"
+    bl_options = {'INTERNAL'}
 
     @classmethod
     def poll(cls, context):
-        return bpy.context.selected_objects
+        return context.selected_objects
 
     def execute(self, context):
         object_list = context.scene.TextureBake_Props.object_list
-        for obj in bpy.context.selected_objects:
+        for obj in context.selected_objects:
             if obj.type != "MESH":
                 continue
             if [i for i in object_list if i.obj.name == obj.name]:
                 continue
             new_item = object_list.add()
             new_item.obj = obj
-        return{'FINISHED'}
+        return {'FINISHED'}
 
 
 class TEXTUREBAKE_OT_remove_object(Operator):
     """Remove the selected object from the bake list."""
     bl_idname = "texture_bake.remove_object"
     bl_label = "Remove an object from the bake list"
+    bl_options = {'INTERNAL'}
 
     @classmethod
     def poll(cls, context):
@@ -712,13 +714,14 @@ class TEXTUREBAKE_OT_remove_object(Operator):
         index = context.scene.TextureBake_Props.object_list_index
         object_list.remove(index)
         context.scene.TextureBake_Props.object_list_index = min(max(0, index - 1), len(object_list) - 1)
-        return{'FINISHED'}
+        return {'FINISHED'}
 
 
 class TEXTUREBAKE_OT_clear_objects(Operator):
     """Clear the object list"""
     bl_idname = "texture_bake.clear_objects"
     bl_label = "Removes all objects from the bake list"
+    bl_options = {'INTERNAL'}
 
     @classmethod
     def poll(cls, context):
@@ -726,13 +729,14 @@ class TEXTUREBAKE_OT_clear_objects(Operator):
 
     def execute(self, context):
         context.scene.TextureBake_Props.object_list.clear()
-        return{'FINISHED'}
+        return {'FINISHED'}
 
 
 class TEXTUREBAKE_OT_move_object(Operator):
     """Move an object in the list."""
     bl_idname = "texture_bake.move_object"
     bl_label = "Move an object in the bake list"
+    bl_options = {'INTERNAL'}
 
     direction: bpy.props.EnumProperty(items=(('UP', 'Up', ""), ('DOWN', 'Down', "")))
 
@@ -744,9 +748,9 @@ class TEXTUREBAKE_OT_move_object(Operator):
         object_list = context.scene.TextureBake_Props.object_list
         old_index = context.scene.TextureBake_Props.object_list_index
         new_index = old_index + (-1 if self.direction == 'UP' else 1)
-        max_index = len(bpy.context.scene.TextureBake_Props.object_list) - 1
+        max_index = len(context.scene.TextureBake_Props.object_list) - 1
         object_list.move(old_index, new_index)
-        bpy.context.scene.TextureBake_Props.object_list_index = max(0, min(new_index, max_index))
+        context.scene.TextureBake_Props.object_list_index = max(0, min(new_index, max_index))
         return{'FINISHED'}